Calculation of order total with request to currency service

+++ order-api/order/model.go
@@ -9,7 +9,7 @@ type Order struct {
 	ID          uint64     `json:"id"`
 	CustomerID  uuid.UUID  `json:"customer_id"`
 	LineItems   []LineItem `json:"line_items"`
-	Total       float64    `json:"total"`
+	Total       float32    `json:"total"`
 	CreatedAt   *time.Time `json:"created_at"`
 	ShippedAt   *time.Time `json:"shipped_at"`
 	CompletedAt *time.Time `json:"completed_at"`
@@ -19,5 +19,5 @@ type Order struct {
 type LineItem struct {
 	ID       uuid.UUID `json:"id"`
 	Quantity int       `json:"quantity"`
-	Price    float64   `json:"price"`
+	Price    float32   `json:"price"`
 }
